@@ -1,18 +1,18 @@
 {
     "require": {
-        "php": "^7.3",
-        "keboola/php-component": "^8.0",
-        "keboola/php-csv-db-import": "^5.1",
+        "php": "^8.1",
+        "keboola/db-adapter-snowflake": "^1.5",
+        "keboola/php-component": "^9.0",
         "keboola/retry": "^0.5.0"
     },
     "require-dev": {
-        "php-parallel-lint/php-parallel-lint": "^1.0",
-        "keboola/coding-standard": "^9.0",
+        "keboola/coding-standard": "^15.0",
+        "keboola/datadir-tests": "^5.0",
         "keboola/php-temp": "^2.0",
+        "php-parallel-lint/php-parallel-lint": "^1.0",
         "phpstan/phpstan": "^0.12.0",
         "phpunit/phpunit": "^9.0",
-        "symfony/process": "^5.0",
-        "keboola/datadir-tests": "^5.0"
+        "symfony/process": "^5.0"
     },
     "autoload": {
         "psr-4": {
@@ -48,6 +48,9 @@
     },
     "config": {
         "sort-packages": true,
-        "optimize-autoloader": true
+        "optimize-autoloader": true,
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true
+        }
     }
 }
